Search Service/requirements.txt
fastapi
uvicorn
elasticsearch
python-dotenv